requirements.txt
pydicom
pynetdicom
requests
pyyaml
pika